refactoring transformers file
