+++ x-pack/legacy/plugins/apm/public/components/app/TransactionDetails/WaterfallWithSummmary/PercentOfParent.tsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import { i18n } from '@kbn/i18n';
 import { EuiToolTip } from '@elastic/eui';
-import { asPercent } from '../../../../utils/formatters';
+import { asPercent } from '../../../../utils/duration_formatters';
 
 interface PercentOfParentProps {
   duration: number;
