+++ x-pack/legacy/plugins/apm/public/components/shared/charts/Timeline/TimelineAxis.js
@@ -12,7 +12,7 @@ import { XYPlot, XAxis } from 'react-vis';
 import LastTickValue from './LastTickValue';
 import AgentMarker from './AgentMarker';
 import { px } from '../../../../style/variables';
-import { getDurationFormatter } from '../../../../utils/formatters';
+import { getDurationFormatter } from '../../../../utils/duration_formatters';
 import theme from '@elastic/eui/dist/eui_theme_light.json';
 
 // Remove any tick that is too close to topTraceDuration
